requirements.txt
fastapi
uvicorn
httpx
aio-pika
sqlalchemy[asyncio]
asyncpg
pydantic
pydantic-settings
PyYAML
openai
alembic
psycopg2-binary
python-dotenv